web/workspace/requirements.txt
Flask
requests
PyMySQL
apache-skywalking==1.0.0